refactor: move logic display in children component
